// cinepath-web/src/app/types/request-types.ts
export interface Keyword {
  id?: number
  name?: string
}

export interface ReleaseDate {
  certification?: string
  iso_639_1?: string
  release_date?: string
  type?: number
  note?: string
}

export interface Video {
  id?: string
  iso_639_1?: string
  iso_3166_1?: string
  key?: string
  name?: string
  official?: boolean
  published_at?: string
  site?: string
  size?: 360 | 480 | 720 | 1080
  type?: 'Trailer' | 'Teaser' | 'Clip' | 'Featurette' | 'Behind the Scenes' | 'Bloopers'
}
// ...
export interface Cast {
  adult?: boolean
  cast_id?: number
  character?: string
  credit_id?: string
  gender?: number | null
  id?: number
  known_for_department?: string
  name?: string
  order?: number
  original_name?: string
  popularity?: number
  profile_path?: string | null
}

export interface Crew {
  adult?: boolean
  credit_id?: string
  department?: string
  gender?: number | null
  id?: number
  known_for_department?: string
  job?: string
  name?: string
  original_name?: string
  popularity?: number
  profile_path?: string | null
}
// ...
export interface PaginatedResponse extends Response {
  page?: number
  total_results?: number
  total_pages?: number
}
// ...
export interface MovieResponse extends Response {
  adult?: boolean
  backdrop_path?: string
  belongs_to_collection?: object
  budget?: number
  genres?: Array<Genre>
  homepage?: string
  id?: number
  imdb_id?: string
  original_language?: string
  original_title?: string
  overview?: string
  popularity?: number
  poster_path?: string
  production_companies?: Array<ProductionCompany>
  production_countries?: Array<ProductionCountry>
  release_date?: string
  revenue?: number
  runtime?: number
  spoken_languages?: Array<SpokenLanguage>
  status?: 'Rumored' | 'Planned' | 'In Production' | 'Post Production' | 'Released' | 'Canceled'
  tagline?: string
  title?: string
  video?: boolean
  vote_average?: number
  vote_count?: number
}
// ...
export interface CreditsResponse extends Response {
  id?: number
  cast?: Array<Cast>
  crew?: Array<Crew>
}
// ...
export interface MovieKeywordResponse extends Response {
  id?: number
  keywords?: Array<Keyword>
}

export interface MovieReleaseDatesResponse extends Response {
  id?: number
  results?: Array<{
    iso_3166_1?: string
    release_dates?: Array<ReleaseDate>
  }>
}

export interface VideosResponse extends Response {
  id?: number
  results?: Array<Video>
}
// ...
export interface MovieReviewsResponse extends PaginatedResponse {
  results?: Array<Review>
}
// ...
export interface Review {
  id?: string
  author?: string
  content?: string
  iso_639_1?: string
  media_id?: number
  media_title?: string
  media_type?: string
  url?: string
}